+++ joulescope_ui/widgets/waveform/marker.py
@@ -327,7 +327,7 @@ class Marker(pg.GraphicsObject):
         self.start_pos = self.get_pos()
         self.setZValue(Z_MARKER_MOVING)
         # https://doc.qt.io/qt-5/qt.html#KeyboardModifier-enum
-        if int(QtGui.Qt.ControlModifier & ev.modifiers()) and self.pair is not None:
+        if int(QtCore.Qt.ControlModifier & ev.modifiers()) and self.pair is not None:
             self.pair.moving = True
             self.pair.moving_offset = self.pair.get_pos() - self.get_pos()
         if self.pair is not None: